+++ apps/docs/components/footer.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import dynamic from "next/dynamic";
+import {usePathname} from "next/navigation";
 
 import {siteConfig} from "@/config/site";
 
@@ -8,6 +9,12 @@ const Link = dynamic(() => import("@nextui-org/react").then((mod) => mod.Link));
 const VercelCallout = dynamic(() => import("./vercel-callout").then((mod) => mod.VercelCallout));
 
 export const Footer = () => {
+  const pathname = usePathname();
+
+  if (pathname.includes("/examples")) {
+    return null;
+  }
+
   return (
     <footer className="container mx-auto max-w-7xl pb-12 px-12">
       <div className="flex flex-col justify-center items-center gap-1">