+++ apps/docs/components/marketing/features-grid.tsx
@@ -9,7 +9,7 @@ import {LinkIcon} from "@nextui-org/shared-icons";
 const styles = tv({
   slots: {
     base: "grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4",
-    card: "dark:border-transparent bg-white/5 dark:bg-default-400/10 backdrop-blur-lg backdrop-saturate-[1.8]",
+    card: "border-transparent bg-white/5 dark:bg-default-400/10 backdrop-blur-lg backdrop-saturate-[1.8]",
     header: "gap-2 pb-0",
     body: "",
     iconWrapper: