+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nextui-org/popover
 
+## 0.0.0-dev-v2-20230625233448
+
+### Patch Changes
+
+- Navbar styles fixed
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230625233448
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230625233448
+  - @nextui-org/use-aria-button@0.0.0-dev-v2-20230625233448
+  - @nextui-org/react-utils@0.0.0-dev-v2-20230625233448
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230625233448
+  - @nextui-org/button@0.0.0-dev-v2-20230625233448
+  - @nextui-org/system@0.0.0-dev-v2-20230625233448
+  - @nextui-org/theme@0.0.0-dev-v2-20230625233448
+
 ## 0.0.0-dev-v2-20230625145649
 
 ### Patch Changes