+++ packages/utilities/react-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nextui-org/react-utils
 
+## 0.0.0-dev-v2-20230625233448
+
+### Patch Changes
+
+- Navbar styles fixed
+- Updated dependencies
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230625233448
+
 ## 0.0.0-dev-v2-20230625145649
 
 ### Patch Changes